fix: fix file-type
