+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "build:ci": "node scripts/build-ci.js",
+    "build:ci": "node scripts/build-ci.mjs",
     "lint": "eslint .",
     "preview": "vite preview"
   },
